+++ src/module/entities/SwadeActor.ts
@@ -381,47 +381,6 @@ export default class SwadeActor extends Actor {
   }
 
   /**
-   * @deprecated
-   * Launches a dialog to configure which initiative-modifying edges/hindrances the character has
-   */
-  async configureInitiative() {
-    const initData = this.data.data.initiative;
-    const template = 'systems/swade/templates/initiative/configure-init.html';
-    const html = await renderTemplate(template, initData);
-    const d = new Dialog({
-      title: 'Configure Initiative',
-      content: html,
-      buttons: {
-        ok: {
-          icon: '<i class="fas fa-check"></i>',
-          label: game.i18n.localize('SWADE.Ok'),
-          callback: async (html: JQuery<HTMLElement>) => {
-            await this.update({
-              'data.initiative': {
-                hasLevelHeaded: (html as JQuery)
-                  .find('#hasLevelHeaded')
-                  .is(':checked'),
-                hasImpLevelHeaded: (html as JQuery)
-                  .find('#hasImpLevelHeaded')
-                  .is(':checked'),
-                hasHesitant: (html as JQuery)
-                  .find('#hasHesitant')
-                  .is(':checked'),
-              },
-            });
-          },
-        },
-        cancel: {
-          icon: '<i class="fas fa-times"></i>',
-          label: game.i18n.localize('SWADE.Cancel'),
-        },
-      },
-      default: 'cancel',
-    });
-    d.render(true);
-  }
-
-  /**
    * Calculates the correct armor value based on SWADE v5.5 and returns that value
    */
   calcArmor(): number {